api/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
flask_apscheduler==1.13.1
pandas==2.2.1
numpy==1.26.4
Flask_Cors==4.0.0
fuzz==0.1.1
fuzzywuzzy==0.18.0
openai==1.30.1